gated_linear_networks/requirements.txt
absl-py==0.10.0
aiohttp==3.6.2
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.2.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
chex==0.0.2
cloudpickle==1.6.0
decorator==4.4.2
dill==0.3.2
dm-env==1.2
dm-haiku==0.0.2
dm-tree==0.1.5
future==0.18.2
gast==0.3.3
google-auth==1.22.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
jax==0.2.0
jaxlib==0.1.55
Keras-Preprocessing==1.1.2
Markdown==3.2.2
multidict==4.7.6
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
promise==2.3
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.24.0
requests-oauthlib==1.3.0
rlax==0.0.2
rsa==4.6
scipy==1.5.2
six==1.15.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-datasets==3.2.1
tensorflow-estimator==2.3.0
tensorflow-metadata==0.24.0
tensorflow-probability==0.11.1
termcolor==1.1.0
toolz==0.11.1
tqdm==4.50.0
